pyproject.toml
[project]
name = "crm-contable"
version = "0.1.0"
description = "Sistema contable para llevar control de gastos"
authors = [
    {name = "Carlos Catarino ",email = "carlos_cata27@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair (==5.5.0)",
    "attrs (==25.3.0)",
    "blinker (==1.9.0)",
    "cachetools (==6.1.0)",
    "certifi (==2025.6.15)",
    "charset-normalizer (==3.4.2)",
    "click (==8.2.1)",
    "gitdb (==4.0.12)",
    "gitpython (==3.1.44)",
    "idna (==3.10)",
    "jinja2 (==3.1.6)",
    "jsonschema (==4.24.0)",
    "jsonschema-specifications (==2025.4.1)",
    "markupsafe (==3.0.2)",
    "narwhals (==1.43.1)",
    "numpy (==2.3.0)",
    "packaging (==25.0)",
    "pandas (==2.3.0)",
    "pillow (==11.2.1)",
    "protobuf (==6.31.1)",
    "psycopg2-binary (==2.9.10)",
    "pyarrow (==20.0.0)",
    "pydeck (==0.9.1)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.1.0)",
    "pytz (==2025.2)",
    "referencing (==0.36.2)",
    "requests (==2.32.4)",
    "rpds-py (==0.25.1)",
    "six (==1.17.0)",
    "smmap (==5.0.2)",
    "sqlalchemy (==2.0.44)",
    "streamlit (==1.46.0)",
    "tenacity (==9.1.2)",
    "toml (==0.10.2)",
    "tornado (==6.5.1)",
    "typing-extensions (==4.14.0)",
    "tzdata (==2025.2)",
    "urllib3 (==2.5.0)",
    "watchdog (==6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
